requirements.txt
aio-pika==9.2.3
aiormq==6.7.7
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
fastapi==0.103.1
flake8==6.1.0
flake8-isort==6.1.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
multidict==6.0.4
packaging==23.1
pamqp==3.2.1
pika==1.3.2
pluggy==1.3.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==3.1.0
pytest==7.4.2
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
yarl==1.9.2
